[modem] chore: move demodulator to own module
